feat: separate fixtures module, rename tomodachi_image fixture to testcontainers_docker_image
