+++ src/tomodachi_testcontainers/pytest/__init__.py
@@ -2,7 +2,8 @@ import contextlib
 
 import pytest
 
-from tomodachi_testcontainers.pytest.localstack_fixtures import (
+from tomodachi_testcontainers.pytest.fixtures.containers import testcontainers_docker_image
+from tomodachi_testcontainers.pytest.fixtures.localstack import (
     _restart_localstack_container_on_teardown,
     localstack_container,
     localstack_dynamodb_client,
@@ -14,7 +15,7 @@ from tomodachi_testcontainers.pytest.localstack_fixtures import (
     localstack_sqs_client,
     localstack_ssm_client,
 )
-from tomodachi_testcontainers.pytest.moto_fixtures import (
+from tomodachi_testcontainers.pytest.fixtures.moto import (
     _reset_moto_container_on_teardown,
     moto_container,
     moto_dynamodb_client,
@@ -26,10 +27,9 @@ from tomodachi_testcontainers.pytest.moto_fixtures import (
     moto_sqs_client,
     moto_ssm_client,
 )
-from tomodachi_testcontainers.pytest.tomodachi_fixtures import tomodachi_image
 
 with contextlib.suppress(ModuleNotFoundError):
-    from tomodachi_testcontainers.pytest.sftp_fixtures import sftp_container, userpass_sftp_client, userssh_sftp_client
+    from tomodachi_testcontainers.pytest.fixtures.sftp import sftp_container, userpass_sftp_client, userssh_sftp_client
 
 
 pytest.register_assert_rewrite("tomodachi_testcontainers.pytest.assertions")
@@ -57,7 +57,7 @@ __all__ = [
     "moto_sqs_client",
     "moto_ssm_client",
     "sftp_container",
-    "tomodachi_image",
+    "testcontainers_docker_image",
     "userpass_sftp_client",
     "userssh_sftp_client",
 ]
