+++ src/tomodachi_testcontainers/pytest/fixtures/containers.py
@@ -10,7 +10,7 @@ from tomodachi_testcontainers.utils import get_docker_image
 
 
 @pytest.fixture(scope="session")
-def tomodachi_image() -> Generator[DockerImage, None, None]:
+def testcontainers_docker_image() -> Generator[DockerImage, None, None]:
     if image_id := os.getenv("TOMODACHI_TESTCONTAINER_IMAGE_ID"):
         yield get_docker_image(image_id)
     else:
